+++ src/apiflask/__init__.py
@@ -1,7 +1,7 @@
-from marshmallow import post_dump as after_dump
-from marshmallow import post_load as after_load
-from marshmallow import pre_dump as before_dump
-from marshmallow import pre_load as before_load
+from marshmallow import post_dump as post_dump
+from marshmallow import post_load as post_load
+from marshmallow import pre_dump as pre_dump
+from marshmallow import pre_load as pre_load
 from marshmallow import validates as validate
 from marshmallow import validates_schema as validate_schema
 from marshmallow import ValidationError as ValidationError