+++ tests/test_decorator_output.py
@@ -2,12 +2,11 @@ from flask import make_response
 from flask.views import MethodView
 from openapi_spec_validator import validate_spec
 
+from .schemas import FooSchema
+from .schemas import QuerySchema
 from apiflask import input
 from apiflask import output
 from apiflask.fields import String
-
-from .schemas import FooSchema
-from .schemas import QuerySchema
 
 
 def test_output(app, client):