+++ tests/test_openapi_security.py
@@ -1,9 +1,9 @@
 import pytest
 from openapi_spec_validator import validate_spec
 
+from apiflask import auth_required
 from apiflask import HTTPBasicAuth
 from apiflask import HTTPTokenAuth
-from apiflask import auth_required
 
 
 def test_httpbasicauth_security_scheme(app, client):